Rename internal/clients package to internal/lib
